xplit xml and json serializer/deserializer
