+++ src/Serialize/XmlSerializer.php
@@ -19,7 +19,7 @@
  * Copyright (c) Steve Springett. All Rights Reserved.
  */
 
-namespace CycloneDX\BomFile;
+namespace CycloneDX\Serialize;
 
 use CycloneDX\Helpers\SimpleDomTrait;
 use CycloneDX\Models\Bom;
@@ -28,13 +28,11 @@ use CycloneDX\Models\License;
 use CycloneDX\Specs\Spec10;
 use CycloneDX\Specs\Spec11;
 use CycloneDX\Specs\Spec12;
-use CycloneDX\Specs\SpecInterface;
 use DomainException;
 use DOMDocument;
 use DOMElement;
 use DOMException;
 use Generator;
-use InvalidArgumentException;
 use RuntimeException;
 
 /**
@@ -42,27 +40,11 @@ use RuntimeException;
  *
  * @author jkowalleck
  */
-class Xml extends AbstractFile
+class XmlSerializer extends AbstractSerialize implements SerializerInterface
 {
     use SimpleDomTrait;
 
-    /**
-     * @var string
-     */
-    private $namespaceUrl;
-
-    public function getNamespaceUrl(): string
-    {
-        return $this->namespaceUrl;
-    }
-
-    public function __construct(SpecInterface $spec)
-    {
-        parent::__construct($spec);
-        $this->namespaceUrl = 'http://cyclonedx.org/schema/bom/'.$spec->getVersion();
-    }
-
-    // region Serialize
+    // region SerializerInterface
 
     /**
      * @throws DOMException
@@ -95,7 +77,7 @@ class Xml extends AbstractFile
     public function bomToDom(DOMDocument $document, Bom $bom): DOMElement
     {
         $element = $document->createElementNS(
-            $this->getNamespaceUrl(),
+            'http://cyclonedx.org/schema/bom/'.$this->spec->getVersion(),
             'bom'
         );
         $this->simpleDomSetAttributes($element, [
@@ -206,152 +188,5 @@ class Xml extends AbstractFile
         return $element;
     }
 
-    // endregion Serialize
-
-    // region Deserialize
-
-    /**
-     * @throws InvalidArgumentException
-     */
-    public function deserialize(string $data): Bom
-    {
-        $dom = new DOMDocument();
-        // @TODO add NOBLANKS ? see if all tests still pass
-        $options = LIBXML_NOCDATA | LIBXML_NOBLANKS | LIBXML_NONET;
-        if (defined('LIBXML_COMPACT')) {
-            $options |= LIBXML_COMPACT;
-        }
-        if (defined('LIBXML_PARSEHUGE')) {
-            $options |= LIBXML_PARSEHUGE;
-        }
-        $loaded = $dom->loadXML($data, $options);
-        if (false === $loaded || null === $dom->documentElement) {
-            throw new InvalidArgumentException('does not deserialize to expected structure');
-        }
-
-        // @TODO normalize
-
-        return $this->bomFromDom($dom->documentElement);
-    }
-
-    public function bomFromDom(DOMElement $element): Bom
-    {
-        $bom = new Bom();
-        $bom->setVersion((int) $this->simpleDomGetAttribute('version', $element, '1'));
-        foreach ($this->simpleDomGetChildElements($element) as $childElement) {
-            if ('components' === $childElement->tagName) {
-                $bom->setComponents(array_map(
-                    [$this, 'componentFromDom'],
-                    iterator_to_array($this->simpleDomGetChildElements($childElement))
-                ));
-            }
-        }
-
-        return $bom;
-    }
-
-    public function componentFromDom(DOMElement $element): Component
-    {
-        $name = $version = null; // essentials
-        $group = $description = $licenses = $hashes = null; // non-essentials
-        foreach ($this->simpleDomGetChildElements($element) as $childElement) {
-            switch ($childElement->nodeName) {
-                case 'name':
-                    $name = $childElement->nodeValue;
-                    break;
-                case 'version':
-                    $version = $childElement->nodeValue;
-                    break;
-                case 'group':
-                    $group = $childElement->nodeValue;
-                    break;
-                case 'description':
-                    $description = $childElement->nodeValue;
-                    break;
-                case 'licenses':
-                    $licenses = iterator_to_array($this->licensesFromDom($childElement));
-                    break;
-                case 'hashes':
-                    $hashes = iterator_to_array($this->hashesFromDom($childElement));
-                    break;
-            }
-        }
-
-        // asserted by SCHEMA
-        $type = $element->getAttribute('type');
-        assert(null !== $name);
-        assert(null !== $version);
-
-        return (new Component($type, $name, $version))
-            ->setGroup($group)
-            ->setDescription($description)
-            ->setLicenses($licenses ?? [])
-            ->setHashes($hashes ?? []);
-    }
-
-    /**
-     * @return Generator<License>
-     */
-    public function licensesFromDom(DOMElement $element): Generator
-    {
-        foreach ($this->simpleDomGetChildElements($element) as $childElement) {
-            switch ($childElement->nodeName) {
-                case 'license':
-                    yield $this->licenseFromDom($childElement);
-                    break;
-                case 'expression':
-                    if ($this->spec->getVersion() >= '1.2') {
-                        // @TOD implement a model for LicenseExpression
-                        yield new License($element->nodeValue);
-                    } else {
-                        trigger_error('Found unsupported LicenseExpression. Using License instead', E_USER_NOTICE);
-                        yield new License($element->nodeValue);
-                    }
-                    break;
-            }
-        }
-    }
-
-    public function licenseFromDom(DOMElement $element): License
-    {
-        $nameOrId = null; // essentials
-        $url = null; // non-essentials
-        foreach ($this->simpleDomGetChildElements($element) as $childElement) {
-            switch ($childElement->nodeName) {
-                case 'name':
-                case 'id':
-                    $nameOrId = $childElement->nodeValue;
-                    break;
-                case 'url':
-                    $url = $childElement->nodeValue;
-                    break;
-            }
-        }
-
-        // asserted by SCHEMA
-        assert(null !== $nameOrId);
-
-        return (new License($nameOrId))
-            ->setUrl($url);
-    }
-
-    /**
-     * @return Generator<string, string>
-     */
-    public function hashesFromDom(DOMElement $element): Generator
-    {
-        foreach ($this->simpleDomGetChildElements($element) as $childElement) {
-            yield from $this->hashFromDom($childElement);
-        }
-    }
-
-    /**
-     * @return Generator<string, string>
-     */
-    public function hashFromDom(DOMElement $element): Generator
-    {
-        yield $element->getAttribute('alg') => $element->nodeValue;
-    }
-
-    // endregion Deserialize
+    // endregion SerializerInterface
 }
